// frontend/src/app/models/ticketAndUserAgent.ts
export interface TicketAndUserAgent {
    ticketId: number;
    userId: number;
    agentId: number;
    ticketType: string;
    createdDate: string;
    ticketStatus: string;
    solvedDate: string;
    reqName: string;
    reqMobile: string;
    reqEmail: string;
    reason: string;
    agentName: string;

    userName: string;
    severity: number;
    openedDate: string;
    remarks: string;
    userMobile: string;
    userEmail: string;
    prevName: string;
    prevMobile: string;
    accountStatus: string;
}